unsigned int -> unsigned long
git-svn-id: svn://orca.physics.unc.edu/Orca64@11890 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/USB/DT5725/ORDT5725Model.h
@@ -225,8 +225,8 @@ typedef struct  {
 - (BOOL)            trigOverlapEnabled;
 - (void)            setTrigOverlapEnabled:(BOOL)aTrigOverlapEnabled;
 //------------------------------
-- (int)				eventSize;
-- (void)			setEventSize:(int)aEventSize;
+- (unsigned long)   eventSize;
+- (void)			setEventSize:(unsigned long)aEventSize;
 //------------------------------
 - (BOOL)            clockSource;
 - (void)            setClockSource:(BOOL)aClockSource;
@@ -384,8 +384,8 @@ typedef struct  {
 - (void) encodeWithCoder:(NSCoder*)encoder;
 
 #pragma mark ***HW Read/Write API
-- (int)     writeLongBlock:(unsigned long*) writeValue atAddress:(unsigned int) vmeAddress;
-- (int)     readLongBlock:(unsigned long*)  readValue atAddress:(unsigned int) vmeAddress;
+- (int)     writeLongBlock:(unsigned long*) writeValue atAddress:(unsigned long) vmeAddress;
+- (int)     readLongBlock:(unsigned long*)  readValue atAddress:(unsigned long) vmeAddress;
 - (void)    writeChan:(unsigned short)chan reg:(unsigned short) pReg sendValue:(unsigned long) pValue;
 - (void)    readChan:(unsigned short)chan reg:(unsigned short) pReg returnValue:(unsigned long*) pValue;
 - (int)     readFifo:(char*)destBuff numBytesToRead:(unsigned long)    numBytes;
